Navbar Cleanup
I removed the white background from the dropdown container. Along with that I assured that there isn't any more white space whenever you hover above the homebrew button. The final touch was the fact that whenver you hover over the current active webpages button you'll find that there isn't the white box-shadow that was there previously, allowing for the buttons to look far cleaner than they were previously. This is the first comprehensive comment that I've ever added and I think that I'll be continuing along this path of advanced and fleshed out Commit comments in the near future.